requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
Click==7.0
colorama==0.4.1
Django==2.2.4
django-cors-headers==3.0.2
djangorestframework==3.10.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
wrapt==1.11.2
